refector: 기존 event handler의 위치를 EventsHandlerModule의 service로 옮김
