--- src/conversation-events/services/conversation-events-logger.service.ts
@@ -1,11 +1,0 @@
-import { Injectable, Logger } from '@nestjs/common';
-
-@Injectable()
-export class ConversationEventsLoggerService {
-  private readonly logger: Logger = new Logger('RoomEventGateway');
-  constructor() {}
-
-  getLogger(): Logger {
-    return this.logger;
-  }
-}
